Fixed outher tangent points calculation

+++ src/utils.cpp
@@ -19,7 +19,7 @@ float degreeMod(float degree, float x)
 
 float calcDistance(float x1, float y1, float x2, float y2)
 {
-    return sqrt(pow(x2 - x1, 2) + pow(y2 - y1, 2));
+    return sqrt(powf(x2 - x1, 2) + powf(y2 - y1, 2));
 }
 
 sf::Vector2f getCircleCenterPos(const Rect& rect, float radius, int moveType)
